Move companion-resolvable extensions into their companion objects
Top-level extensions on a module's own concrete types resolve via the
receiver's implicit scope when placed in the companion, so they need no
import and don't clutter the package namespace. Move Diff, Tel and Cursor
extensions accordingly, and document the resolution rules in syntax.md.
