+++ lib/stratiform/src/core/stratiform.Tel.scala
@@ -61,6 +61,28 @@ object Tel extends Tel2:
   type Error = TelError
   val Error: TelError.type = TelError
 
+  extension (tel: Tel)
+    def edited(revision: Revision): Tel raises MutationError = revision(tel)
+
+    // Runtime-checks `tel` against `topic`, then re-types it as a schema-typed
+    // `Tel of topic from topic`. The phantom `Topic` records the current position
+    // within the schema (initially the root, `topic`) and `Origin` records the
+    // root schema, so the `Dynamic` navigation macros can resolve fields and
+    // child indices against `topic` at compile time — with no `dynamicTelAccess`
+    // import required.
+    //
+    // The runtime conformance check is a decode-and-discard against `topic`'s
+    // `Tel.Decodable` (a nonconformant value raises `TelError`): TEL scalars are
+    // untyped text, so a structural walk over the schema cannot catch type mismatches
+    // (e.g. a non-numeric `Int` field) that decoding does — hence decoding remains the
+    // strict check. The carrying `Tel.Decodable` also supplies the schema for the
+    // phantom anchor, so the schema is guaranteed coherent with the decoder used.
+    def verify[topic](using topic is Tel.Decodable)
+    :   (Tel of topic from topic) raises TelError =
+
+      tel.as[topic]
+      tel.asInstanceOf[Tel of topic from topic]
+
   object Encodable:
     def apply[value](shape0: => Morphology)(lambda: value => Tel): value is Tel.Encodable =
       new Tel.Encodable:
